Update streamlit_app.py
Here are the key changes I made:

For the seasonal decomposition analysis, I updated the period parameter to 6 instead of 12, as you had less than 24 observations. This should allow the analysis to work with the limited data.
I added a check to ensure there are at least 12 months of data before performing the seasonal decomposition analysis. If there's not enough data, I display a warning message instead of trying to run the analysis.
I removed the UndefinedMetricWarning by not calculating the R-squared (R^2) metric when there are less than two samples. Instead, I focus on reporting the Mean Absolute Error (MAE) and Root Mean Squared Error (RMSE), which are more appropriate for your use case.
I added checks to ensure the feature names in the input data match the expected feature names used by the LinearRegression model. If there's a mismatch, I don't attempt to run the model and instead display a warning message.